Complete refactor.
Clean up instantiation code and simplify the wiring abstraction. We now
use a 2D matrix to store the interior nodes so that we can always deal
with (level, index) instead of translating between that form and a
1D array index form.

Additionally, we no longer store references to nodes as inputs since
that was going to make mutation tricky. Now we just store a
(level, index) tuple for each input. This means that a node can be
replaced without any re-wiring.

Code has been split into fairly logical files for simplicity and sanity.

Still need to write tests for some of the helper functions and verify
that the random instantiation actually works properly.